created NavContent comp which conditionally renders based on login true/false
